requirements.txt
aiohttp
bs4
discord_webhook
pyyaml
python-dateutil
pytz
lxml
